CHashContext.cpp:   Fix wrong array indexing CHashContext.h:   Revise parameter naming. HashEdDlg.cpp:   Add null pointer protection. PhraseList.cpp:   Align listbox's selector with it's data.
git-svn-id: https://svn.csie.net/chewing/win32-chewing@691 ac1be623-90ea-0310-9410-ba68b2efa777

+++ HashEd-UTF8/HashEdDlg.cpp
@@ -429,6 +429,9 @@ void CHashEdDlg::OnDelPhrase()
 	if ( selItem==-1 )  return;
 
 	pItem = (HASH_ITEM*)phraseList->getItemData( selItem );
+	if (pItem == NULL) {
+		return;
+	}
 	phraseList->deleteItem(selItem);
 
     m_context.del_phrase_by_id(pItem->item_index);
